# release-20.2: opt: reduce allocations in Implicator and JoinOrderBuilder