--- a/pkg/sql/opt/partialidx/implicator.go
+++ b/pkg/sql/opt/partialidx/implicator.go
@@ -145,7 +145,6 @@ func (im *Implicator) Init(f *norm.Factory, md *opt.Metadata, evalCtx *tree.Eval
 	im.f = f
 	im.md = md
 	im.evalCtx = evalCtx
-	im.constraintCache = make(map[opt.ScalarExpr]constraintCacheItem)
 }
 
 // FiltersImplyPredicate attempts to prove that a partial index predicate is
@@ -647,9 +646,18 @@ func (im *Implicator) twoVarComparisonImpliesTwoVarComparison(
 	return false, true
 }
 
+// initConstraintCache initializes the constraintCache field if it has not yet
+// been initialized.
+func (im *Implicator) initConstraintCache() {
+	if im.constraintCache == nil {
+		im.constraintCache = make(map[opt.ScalarExpr]constraintCacheItem)
+	}
+}
+
 // cacheConstraint caches a constraint set and a tight boolean for the given
 // scalar expression.
 func (im *Implicator) cacheConstraint(e opt.ScalarExpr, c *constraint.Set, tight bool) {
+	im.initConstraintCache()
 	if _, ok := im.constraintCache[e]; !ok {
 		im.constraintCache[e] = constraintCacheItem{
 			c:     c,
@@ -662,6 +670,7 @@ func (im *Implicator) cacheConstraint(e opt.ScalarExpr, c *constraint.Set, tight
 // cache contains an entry for the given scalar expression. It returns
 // ok = false if the scalar expression does not exist in the cache.
 func (im *Implicator) fetchConstraint(e opt.ScalarExpr) (_ *constraint.Set, tight bool, ok bool) {
+	im.initConstraintCache()
 	if res, ok := im.constraintCache[e]; ok {
 		return res.c, res.tight, true
 	}